MEN-3970: Add support for probing the U-Boot environment separator
This change is needed to support older U-Boot versions, where the firmware tools
only support the 'key value', syntax, whereas the newer tools, and libubootenv
especially only support the '=' separator, giving 'key=value' as the only option.

This is solved through probing the environment by first writing a dummy variable
using the old 'key<whitespace>value' syntax, and then reading it back to see if
the write was succesful (if the bootloader only supports the 'key=value' syntax,
then no value will be set). Hence if the read is succesful, then the environment
only supports the legacy syntax, and if it is not we just have to trust that it
is 'key=value', as in all newer integrations.

Changelog: Title
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>