MEN-3980: Fix: Roll forward client error
Currently the client has some issues with rollback on a current running
partition. Hence, any error which takes the client to the rollback state,
without the client rebooting, will actually make the client 'roll forward',
meaning that it flips the 'mender_boot_part' partition variable, when it is not
meant to. This leads a client to reboot into the inactive partition.

This fix checks to see if the mounted partition and the next bootable partition
matches. If it does not, then a reboot is not needed, since the client came to
the rollback state without having rebooted, which can only happen on the active
partition after an install error during the initial update process.

In all other cases, the partition will be flipped.

Changelog: Fix: Do not switch boot partitions on installation errors on the
active partition.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit dd11c62ec8ae9e0c932b150db2779994a864e271)