MEN-4032: Add Glib's GIO dependency and demo integration
Added a new package `dbus` in the source code defining an interface and
a default implementation for it (using C bindings for libgio).

Implement wrappers for the two easiest calls and add unit tests for
them.

Changelog: Add Glib's GIO dependency for D-Bus interface. It can be
opt-out using `nodbus` at compile time.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>