Remove Makefile optimization to reuse old binary if source is unchanged.
This does not pick up when TAGS have changed, which is important for
some tests in meta-mender. While it would be possible to fix this on
the Yocto side by cleaning first, Go is quite good at caching builds
these days, so this optimization is not giving much value.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>