[Makefile]: Add a separate install target for network scripts
This way, it can later on be made optional, whether or not to include the
'inventory-geo-script', and other network scripts if we decide to add them.

This is followed by a separate target in meta-mender, called:
mender-inventory-network-scripts, which can be removed ad-hoc by a superuser which
whishes full control over his install.

This follows from the situation where we have had some issues with the
third-party provider we rely on for the geolocation data.

The script is still a part of the standard install target.

Changelog: The 'inventory-geo-script' now has a separate install target:
'install-inventory-network-scripts'. Note however that it is still installed by
the default 'install-inventory-scripts' target.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>