Add: inventory-script for reporting installed update-modules
This simple scripts lists all the executable files found in the inventory-script
path:

/usr/share/mender/modules/*/*

If no update-modules are found, then an empty inventory is set. This is so that
an update, which switches from a setup with update-modules to one completely
without them will reset the inventory that is already reported.

Changelog: Add an inventory script for reporting the update-modules currently
installed on a device.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>