Turn off exit-on-close flag for DBus connections.
The DBus connection is a service to other processes, not a core part
of how the server operates, so it should keep running even if it's
broken.

This manifested as the test binary sometimes terminating abnormally at
the end of the test run, when the DBus test server is terminated.

Changelog: none

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>