Make MenderAuthManager.Stop() synchronous and fix race condition.
Being synchronous is not required for the client since we only expect
to stop the auth manager right before quitting, but it helps in
testing by making the execution flow more predictable. Arguably, it is
also more future proof.

Somewhat independently, but exposed because of the new synchronicity,
a race condition was fixed. Golang does not support volatile
variables, and it is not safe to use the same variable in multiple go
routines. This would sometimes manifest as the `running` variable
evaluating to true even after being set to false. Presumably,
`running` would reside in a register while inside the `Stop` function,
only being written to memory after returning, at which point it's too
late. It was replaced by channels instead.

Also remove the check for closed inChan channel. It is not closed
anywhere in the source file, so it's actually better to get a panic if
this ever happens, so that we can catch it.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>