Fix auth token not being loaded after rebooting to new partition.
This showed up in an integration test which uses the tenant token.
Strictly speaking, it is an invalid test, because it starts on a
partition with a tenant token, and then deploys an image that doesn't
have the token. The reason it has been passing is because the JWT auth
token is loaded, which is still valid even though the tenant token is
not present anymore. This is a bug of its own, which we may address
some day, but for now, it served as a nice way to expose that the JWT
token was not being loaded.

The problem is that the functions that use the token, use the struct
member directly, instead of an access function which can make sure
it's populated if it isn't already. So switch to using GetAuthToken().

And then remove the cached authToken in the Mender client struct. We
should always load the token from the manager, because other processes
that use the DBus API may have changed the token behind our back.

This also means that several test functions that previously called
Authorize, should not call it anymore, because this leads to calling
the auth endpoint. For functions that are not testing this explicitly,
adding the token to the database is enough, which means it will load
it from there without calling any endpoints.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>