Make sure that re-authorization works.
The functions that need to re-authorize are doing so by calling
Authorize(), but this function caches the result, resulting in an
endless loop of "we have a cached token -> use it -> server rejects it
-> reauthorize -> we have a cached token -> use it ....".

Solved by promoting the internal authorize() function to Authorize(),
and removing the version which caches the result. In all places where
this function is called, it is correct to request a new token.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>