Make the auth manager service routine auto-start.
This is not necessary for normal operation, but it helps testing a lot
by avoiding the setup and teardown in a lot of tests. The next commit
removes the caching of the token, which means the only place the
client can get it is via the manager. This increases the number of
tests needing the setup and teardown code even more, hence the
motivation for this change.

Also make a couple of exported functions that don't need to be
exported private.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>