tests: Work around race condition when detecting bad certificate.
If a bad certificate is detected, which is what we are testing, the
Golang request processing (http.Client.Do) will close the connection.
Sometimes, but not always, this will lead to a "connection closed"
error instead of a "bad certificate" error, the latter of which we are
looking for in the string. It appears that this is outside of our
control, since our server is not torn down until after the check has
been made.

Work around this by repeating the test multiple times using
assert.Eventually.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>