MEN-4170: Fix client not terminating correctly on SIGTERM
A signal handler for SIGTERM is added to the daemon run, but this is only able
to interrupt the client on the state-machine changing state. Thus operations
taking a long time, such as writing an update, are never in fact interrupted on
SIGTERM. Also, the Reboot state is stopped, due to this addition.

This can cause long shutdown times on clients in the process of writing an
update, as the client will not terminate until it has either switched state, or
worse, if it is already in the Reboot state, a SIGKILL signal needs to be sent
from the outside in order to shut it down.

This fix simply exits with Exit code 0 upon a received SIGTERM. The client is
already safe from interruptions, so this call is fine.

Changelog: Exit with code 0 on a received SIGTERM signal.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>