Fix body not being resubmitted when retrying requests across servers.
There are two issues at play here: The first is that we need to get a
new body reader each time we make a new request. This is the call to
`GetBody()`. Otherwise the body reader has already been consumed by
time we make the second call.

The second issue, which is much more subtle, is that we need to
specifically assign the `GetBody` function pointer when we make a new
request. By reading the Golang documentation, one might conclude that
this should not be necessary, and in many cases it isn't, because
Golang is supposed to deduce a sensible `GetBody` function, when a
`body` is submitted to `NewRequest` which has a well-known type. But
there is subtle issue with this: The *new* body, which is returned by
the automatically generated `GetBody` function, does not itself
fulfill the criterion of being a well-known type, hence this operation
only works once, and not "recursively". This is because the original
body type is bytes.Buffer, whereas the generated type is bytes.Reader
[1].

Also extend tests considerably to test not only what happens when a
server cannot be contacted, but when it can be contacted and returns
various invalid responses, including Unauthorized responses.

Changelog: None

[1] https://github.com/golang/go/blob/66c02645062561ac29d00297e8d8c49698b2e4da/src/net/http/request.go#L893

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>