Fix DBus race condition between mender-client and mender-shell.
Since we do not have DBus activation enabled in the
`mender-client.service` file, there is no guarantee that mender-client
will grab the DBus interface first. If mender-shell happens to grab it
first, it will not get a reply to GetJwtToken, but will instead start
waiting for a state change. But unless the token has actually changed,
it will not get one if mender-client doesn't post it on startup.

Adding DBus activation in the systemd service file would also solve
the problem, but this comes with some extra complexities because the
client can also be compiled without DBus support. Rather than deal
with that, this short fix seems to solve the problem well, and also
works if the client simply crashes and comes back online.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>