MEN-4518: Always send inventory after a successful deployment
Previously a reboot was needed in order to force-resend the inventory. This was
never discovered, as the client up until quite recently has almost always
rebooted.

The fix is dead simple, reset the in-memory storage of the last
inventory-update, which is equivalent to a reboot.

Changelog: Send the inventory after a successful deployment, even though the
device has not rebooted.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>