MEN-4535: UpdateControlMap DBus endpoint support
This adds a separate DBus endpoint for DBus under:

Connection name: `io.mender.UpdateManager`
Object path: `/io/mender/UpdateManager`
Interface: `io.mender.Update1`
Method name: `SetUpdateControlMap`

The method takes a `JSON` string, with the `ID` and `Priority` of the control map.
Which is then set in the global `ControlMap` structure.

It returns an integer, representing the `refresh_timeout` of the client.

Changelog: MEN-4535: Add a DBus endpoint for the UpdateControlMap, which allows
a user to set the `ID` and `Priority` of a given update process.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>