MEN-4546: Validate and sanitize Update Control Maps passed via DBus
If the validation fails, an error is returned to the caller via DBus
endpoint.

If the sanitization of a particular state results in default values, it
is removed from the map. Then if the sanitized map ends up with 0
states, it is not added to the pool (but no warning is emitted to the
caller).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>