MEN-4669: Device_type_file is now consistent with -d flag & mender.conf
Changelog: The location of the device type file is now determined by the mender.conf file. If the device type file is not used in mender.conf, the device type file is determined by the --data flag and if the flag is not used, the device type file is set to default. In addition, the scripts and modules directories location is consistent with the --data flag now.

Signed-off-by: Nils Olav Kvelvane Johansen <nils.olav@northern.tech>