Fix Update Control Maps having no effect.
The heart of the issue is that the control map pool was being created
twice, and then the state logic and the update logic were working on
different pools. To remedy this, move the initialization of pool into
NewDaemon, which is really where it belongs, since it cannot be used
in non-daemon mode [1]. This allows it to be created once and used
everywhere.

One of the unit tests which tests a nil argument is invalid after this
change, so just remove it.

[1] Actually the same thing can be said for the auth manager, but
    moving it has a lot more ripple effects which I don't want to deal
    with right now.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>