Don't report the same pause status more than once.
The main reason is that the loop will spin every time a map is
changed, even if this change has no effect on the action. If some
overeager process is submitting maps very frequently, this can result
in a lot of API calls, but we only need the first one as long as the
status is the same.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>