Fix use-after-free issue when receiving signal.
When runDaemon() returns, the first thing that happens is that it
cleans up the resources used by the MenderDaemon structure. But if we
got a signal, we escaped runDaemon without finishing Run(), which
means this memory is still in use. This is worse than simply exiting,
because operations can be corrupted, as opposed to simply
truncated. There is no way to make this safe with the current design.

Therefore split the signal handling, so that SIGTERM is handled in
doMain, directly where we want to exit from. This means that runDaemon
can keep running without its data being destroyed.

An alternative would have been to simply call os.Exit() from the old
location, but it seemed better to keep the exit codes in one place.

Changelog: Fix occasional crash when exiting using SIGTERM.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 654e61dac40eddb8d88c4b96b314dd47720ad1f8)