Conditionally compile the client with the go race detector enabled
This adds a make variable `RACE`, which when set, adds the `-race` flag to the
go compiler, instructing it to build the client with the race-detector enabled.

This will produce a much larger binary, so should in no way be part of any
production setup, but can be handy for example, when running the client as a
part of our test suite.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>