Make command output less ugly in the log.
This is a followup to 26f2ccbd523932. After that commit, the output is
printed correctly, but completely without context, making it hard to
follow in logs. Since the primary purpose of having this output is
debugging, wrap it in a writer to prettify the output.

No changelog, since this is just augmenting the mentioned commit.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>