Mark deployment as failed on bad UpdateControlMap instead of retrying.
Make sure that the deployment fails if the Update Control Map is
invalid. This is important for several reasons. First, there is no
validation on the server, so submitting the map by manual API call has
a quite a high chance for mistakes. These are difficult and cumbersome
to detect if the client silently rejects the update and tries again
later. Second, if the Update Control Maps are extended to include more
states, old client would end up in the same trap, retrying forever,
but giving no message of doing so. Better to fail the deployment
properly.

There are two difficulties in doing this. First, we need to return the
UpdateInfo structure, even when returning an error, so that we have
somewhere to report the failure to. Second, since an invalid map will
(or at least can) cause no data to be returned by the JSON, we need to
parse it again if there is an error, but ignore map errors.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>