MEN-3360: Make state trans. counter ignore certain states.
We divide the states into three groups: Update states which are not
allowed to loop, update states which are allowed to loop, and other
states (which are always allowed to loop). For update states
(`updateState`) we use the new `PermitLooping` interface to determine
whether it permits looping, for other states (`baseState`) it's enough
to know that they are not an `updateState`.

The motivation behind all this is to allow going back and forth
between states as long as at least one of them allows looping, which
will allow us to implement limitless state transitions during certain
operations in a deployment. This is important to support Update
Control, and may have other uses in the future as well (the ticket has
some useful links).

However, using the three groups of states, we *also* maintain the
ability to break out of loops between states that do *not* permit
looping, which typically is caused by update failures. Notice in
particular that reboots, due to calling `LoadStateData`, will always
increase the counter by at least one, even if the states involved do
permit looping.

This also reverts eb15ff94bf4ed57c, which increased the maximum number
of state transitions in order to deal with some states that can loop
for a limited time. In principle, the limit for these states could be
raised now, but it's not something I will handle in this commit.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>