MEN-4829: Fix busy-wait bug for only non-server Update Control Maps wait
In an scenario of only non-server Update Control maps, the client was
busy looping between controlMapState (aware of an active map) and
controlMapPauseState (which checked only for maps matching a given ID).

This fix makes controlMapPauseState.Handle return either a Wait(...)
when non-server maps are present or a MultiplexWait(...) when server
maps are present. In both cases, the times for waiting are HalfWay(s).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>