Fix wrong transition to cleanup after artifact header failure.
While I don't think it has any ill effects, it is not the correct
state to go to because no action has been done which we need to clean
up after. It doesn't fit into the other error paths, which all use
updateStatusReportState instead.

After StorePayloads(), the code switches to updateCleanupState errors,
but this is later in the code.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>