MEN-4857: Make sure incorrect zero values are not included in config.
Changelog: Do not put useless and sometimes even incorrect zero values
in the configuration file when running `mender setup`.

Note that there are still a few values that show up in the file
without reason, like `DBus.Enabled`. Since it is set to true by
default, we can't use `omitempty`, since it will do the opposite of
what we want.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>