MEN-5011: Split client and Auth Manager into two separate components.
This commit splits the client so that Auth Manager is in its own
package and runs in a separate Go routine. In future commits it will
also be possible to disable it. The client code now does not use the
Auth Manager directly, but instead just makes sure it runs, and then
makes all queries over D-Bus.

Note that there are still some significant code dependencies from the
client to the auth manager, for these reasons:

* The client must be able to launch the Auth Manager, since we don't
  want it in a separate process for space efficiency reasons.

* `mender setup` needs to handle configuration for both the client and
  the Auth Manager, for backwards compatibility / usability reasons.

* Tests are still making rather heavy use of the Auth Manager. It
  would be possible to rewrite these to use mocks, but this seems
  undesirable, since the testing is much further from reality, and we
  won't get rid of the dependency because of the other two points
  yet. This also means that many tests in the client require a bit
  more elaborate setup now, with both the manager and a D-Bus server
  needing to run.

However, there is no dependency from the Auth Manager to the client,
so this is possible to compile completely separately (but there is no
code for this currently, since we don't need it now).

About the code structure, the code is split into three new main
directories:

* `common` which contains code that both the client and the Auth
  Manager uses. Currently this includes TLS communication, even though
  we want the the client to proxy straight HTTP through the Auth
  Manager later (MEN-5013). However, we still need the dependency
  because of standalone mode, which cannot use the Auth Manager.

* `authmanager` where all the Authentication code now lives.

* `client` where the client code lives. This no longer does
  authentication, but queries over D-Bus for a token.

Another behavior change (other than the Auth changes):

* The JWT token is no longer stored in the database, but is kept in
  memory only. This is to prevent the client from auto-authenticating
  when it has restarted. After restarting it is important to
  re-authenticate, so that we know the current configuration is valid. TODO

Smaller changes and fixes:

* Race condition in AuthManager creation fixed (see `initDone`
  channel, mainly test related).

* No longer re-authenticates needlessly for many API calls (this part
  is completely rewritten).

* All database keys are moved into their own `dbkeys` package to be
  shared, and to be in one central place.

* Authorize state completely removed. It is now handled transparently
  inside the ApiClient and is no longer a separate state. The
  Unauthorized state no longer exists. This state was never visible
  externally (except through logs), so it doesn't change behavior.

* Some places in the client, in particular the `mender setup` code,
  makes use of the `combinedConfig` struct. This struct is not very
  nice, but it was the only way I could think of which combines the
  two configuration structs (client and auth manager), which are
  separate in most locations, except for in this particular place.

* Implemented D-Bus calling API based on Glib.

* Remove godbus completely. We only used it in tests, but it is no
  longer needed since we have both a calling and listening API based
  on Glib now.

Changelog: Client will no longer cache the Authorization token from
the server across restarts, meaning that it is no longer possible to
end up in the situation where a rootfs update with invalid
authorization data succeeds, only to fail authorization later on when
the token expires.

Changelog: It is no longer possible to change the identity of a device
with a rootfs update. This was not supported before either, but worked
in a hacky way by abusing the authorization token to get a rootfs
update with new identity data to succeed. Afterwards the device would
show up as a new device when the token expired.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>