Fix `go vet` cmd after refactor and start ignoring composites warnings.
First off, we never ran `go vet` for the whole project because it was
missing `./...`, so the others errors being fixed here is because we
now do.

Then we fix the error `copies lock value` because a Mutex is being
copied in the full copy we do of the UpdateControlMap. Replace this
with manual copying which treats this field specially.

Then we fix the message "Composite literal uses unkeyed fields" from
`go vet`, by adding `-composites=false`. I couldn't find a lot of
information/recommendations on it online, except this [1]
reference. In this thread they (including Rob Pike) argue that keyed
fields are better because they do not break on struct reorderings or
extensions. Which is what you want in most cases.

But not all cases! Especially the comment by Space A. is interesting:

  I'm using unkeyed fields for the opposite reason: I don't want code
  to compile when I add new fields to the struct, unless this is
  reflected at the other end.

Which is exactly what we want in the UpdateControlMap copy step. This
code should break if we add fields because we want it to be a full
copy. It is not possible to achieve this with keyed fields, so it is
better to use unkeyed fields and disable this warning.

[1] https://groups.google.com/g/golang-nuts/c/Ub1GMMrfgFc

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>