Check that device can re-authenticate before committing.
In case the installation changed the authentication material, such as
the identity.

Changelog: None
No changelog, the essence of this fix is already captured in
d491f720282e6; this commit is just further making it robust by making
it apply to all updates, not just updates that reboot (rootfs
updates).

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>