Make it safe to register multiple DBus objects with Libgio.
It is important to avoid race conditions while testing, since Libgio
works asynchronously, we need to make sure that each registered object
calls back into the dbus object that registered it, not a global one.

At the moment this isn't used, but on the feature-unified_comms
branch, this was used in tests. We decided to abandon that branch, but
I thought it best to rescue this cleanup so that if we do add more
advanced DBus functionality, we won't re-expose this bug.

Original commit: 8c277bb2e12b294a97977c0c224c097dc0442c08

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>