Also obey update poll interval when querying for control map updates.
The update control map expiration can be long to avoid unintended
expiration when network connectivity is spotty, but one might still
want short update poll intervals. Therefore, choose the smaller of the
two as the timeout to enable rapid feedback.

Changelog: When an update is paused, the client now queries the server
for updates using the `UpdatePollIntervalSeconds` interval if it is
shorter than `UpdateControlMapExpirationTimeSeconds`, enabling quicker
response when continuing an update.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>