MEN-5217: Stop caching auth token and rewrite auth code.
* AuthTokenName and AuthTokenCacheInvalidatorName database fields are
  retired. The token is now only kept in memory and is always
  re-requested when restarting. This solves some issues with token
  caching, see the changelog.

* ApiClient refactored into the rewritten ReauthorizingClient. It
  behaves a bit better than the previous implementation. The old code
  would always try the servers in the the listed order, even if it is
  already known that the first server rejects us. This would cause a
  new login for the second server for every API call. Now it stays on
  the server which last admitted us, until we are either rejected or
  it becomes unavailable.

* Authorize state completely removed. It is now handled transparently
  inside the ReauthorizingClient and is no longer a separate
  state. The unauthorized state no longer exists. This state was never
  visible externally (except through logs), so it doesn't change
  behavior.

* Authorization header is no longer sent to the host we download
  artifacts from, which is expected to use presigned links
  instead. This prevents leaking tokens to third party storage
  servers.

* TestFailoverAPICall test is removed because this is already well
  covered in TestApiClientRequest

Changelog: Client will no longer cache the Authorization token from
the server across restarts, meaning that it is no longer possible to
end up in the situation where a rootfs update with invalid
authorization data succeeds, only to fail authorization later on when
the token expires.

Changelog: It is no longer possible to change the identity of a device
with a rootfs update. This was not supported before either, but worked
in a hacky way by abusing the authorization token to get a rootfs
update with new identity data to succeed. Afterwards the device would
show up as a new device when the token expired.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>