Check that device can re-authorize before committing.
In case the installation changed the authorization material, such as
the identity.

Changelog: None
No changelog, the essence of this fix is already captured in
eaa9a2873e2334; this commit is just further making it robust by making
it apply to all updates, not just updates that reboot (rootfs
updates).

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>