Rearrange log messages to give more info after addition of proxy.
After the local http proxy was added, the log message when successful
authorization data has arrived isn't very helpful. The reason is that
it reports the local proxy address, instead of the remote addres. Fix
that by moving this piece of information to the Auth Manager, where
the remote address is known.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>