MEN-5096: Refresh update control maps before failing an update
Fix the issue where if a state takes longer than the expiration time
for the enabled update control map, then by the time the client checks its
control maps, the map is expired, and the update fails.

Now the client refreshes the update maps from the server before each pause, and
hence, this issue will be avoided.

Changelog: commit
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>