Fix: Increment the retry count in fetch control maps
The variable was previously stored only in the struct, and hence did not survive
a spin of the state machine.

The fix is easy, just store it in the state machine context.

In fact, this can be sort of dangerous, and losing connection whilst refreshing
a control map will actually lead to an infinite loop, which the client will not
break out of until network connection is regained.

Changelog: Handle the possibility of losing network connectivity when refreshing
the update control maps.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 0bebe0b1326b7516fec8c90c409eccd2aed2b76a)