fix(client_update): Only fall back to older endpoints on HTTP 404's
Ticket: MEN-5421

Make the client only fall back to older endpoints on HTTP 404 error codes when
polling the `deployments/next` endpoint.

Previous functionality fell back to older endpoints on all error codes. This in
turn meant that when the client got rate-limited by the server on 429's, the
client fell back to the older endpoints.

The problem with this is that only the POSTv2 endpoint supports control maps, so
when an update with control maps was in progress, the other endpoints would
return 204, and the client would think the deployment was aborted from the server.

Changelog: commit
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 2473653161d382708e14b0403e628310382cf4f8)