fix(control-map-retry): reset the control map fetch retry counter
Previously the counter was never reset to zero, except if the client got
restarted. This will manifest itself in long-running systems not rebooting in
any way, as the `Tried maximum number of times` error.

The fix is simple, just reset the counter after the client is done retrying fetching the
control maps.

Changelog: Fixed an issue in which long-running systems, with a long time
between reboots, and multiple updates encountered the `Tried maximum amount of
times` error, due to an error in the retry logic.
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 099c9beffd3162294fba7247a3f8160c950747ce)