fix: ModuleInstaller.callModule() should check the `capture` parameter
When MEN-5098 (#864) was implemented it looks like it inadvertently
stopped checking the value of the `capture` parameter in the
ModuleInstaller.callModule() method. This means that the module output
is always captured in `buf` instead of using the `stdoutLogger` that
was added in that PR.

The proposed fix in this commit changes ModuleInstaller.callModule() so
that it checks the value of the `capture` parameter before replacing
`cmd.Stdout` with a buffer. This then allows the script stdout to be logged
unless the caller explicitly asks to capture the output.

Changelog: Only capture module stdout when requested
Ticket: None
Signed-off-by: Domenic Rodriguez <domenic.rodriguez@geckorobotics.com>
(cherry picked from commit 37f754f4fc4823ec77651cf70a4d0b3688c4dc7c)